requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# XAI Libraries
captum>=0.6.0
grad-cam>=1.4.8
shap>=0.42.0

# Manifold & Clustering
umap-learn>=0.5.3
scikit-learn>=1.3.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Management
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0

# LLM Integration
openai>=1.0.0
transformers>=4.30.0
tiktoken>=0.5.0

# Data & Utils
pandas>=2.0.0
opencv-python>=4.8.0
tqdm>=4.65.0
pyyaml>=6.0

# Uncertainty
uncertainty-toolbox>=0.1.0

# Development
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional - Diffusion for counterfactuals
# diffusers>=0.21.0
# accelerate>=0.21.0
